Implemented PacketManager
the PacketManager can be accessed by Spout.getPacketManager() It can
send SpoutPacket instances and checks if the spout plugin is enabled and
if this is a multiplayer game. Also, there is an additional method,
sendSpoutPacket(SpoutPacket packet, int pluginVersion) which only sends
the packet when the plugin on the server has pluginVersion or above.