Modifier Keys also for KeyBindings
Also, changed the way when the screen for ambigous bindings is opened.
When you have no modifier or just shift pressed, it will show a basic notification, else it will show the screen because you're probably not typing text in.
When Bindings are not ambigous, it will always trigger them, then it comes to the Plugin or Addon to check if the action could be disturbing

Signed-off-by: narrowtux <narrow.m@gmail.com>