Added glColor4f call to renderString
Mods that rely on renderString to leave the drawing colour set to that last used in a string can continue to do so.

Applied colour-code optimization so that it treats a set of "don't cares" (unused codes) to further simplify the logic used. (Although I still think that lowercase codes happen with a far greater frequency than uppercase without solid data I cannot argue either way.)

Changed 'TextAlpha' to 'AlphaText' and fixed typo.

Signed-off-by: dluechoy <dluechoy@gmail.com>