[FIX] Exit singleplayer goes to favorites
Fixed that exitting a singleplayer game would get you to the favorites.
This now also correctly handles the distinction if you joined from the favorites or the server list.
When you exit from singleplayer it will get you to the main menu, as opposed to multiplayer, which brings you back to one of the multiplayer screens.