Removed ping from server list
Fixes #SPOUTCRAFT-394
Client will also not ping random servers any more, as the ping statistics will be removed from the website as well (looking at @Top_Cat).
Removed "order by ping" option from the filters.
Ping will not be shown on servers that are on the server list, but it will still be shown for servers in the favorites.

Also disabled the Texture pack database (it's a mess).