Fix for CustomBlocks being offset incorrectly in ItemFrames.
Purposely did NOT inset them into the ItemFrame due to some custom blocks not having a square dimension.  Some items that are signs etc that extend material Block don't use a block shape, just its collision box.  This method of aligning the image is used so that the CustomBlock is fully displayed without losing half of the image to the image frame itself.

Fixes: https://spoutdev.atlassian.net/browse/LEGACY-111

Signed-off-by: Mike Howe <dockter@almuramc.com>